+++ src/components/BtnClearAll/index.js
@@ -1,0 +1,1 @@
+export { default } from './BtnClearAll';
