+++ src/components/Choices/Choices.js
@@ -4,11 +4,11 @@ import includes from 'lodash/includes';
 import isEqual from 'lodash/isEqual';
 import uniqueId from 'lodash/uniqueId';
 import without from 'lodash/without';
-import { DISPLAY_NAME_FILTER_CHOICES, NAME_PREFIX_CHECKBOX, NAME_PREFIX_RADIO } from '../constants';
-import CheckBox from './CheckBox';
-import Caption from './Caption';
-import FilterInfo from './FilterInfo';
-import RadioButton from './RadioButton';
+import { DISPLAY_NAME_FILTER_CHOICES, NAME_PREFIX_CHECKBOX, NAME_PREFIX_RADIO } from '../../constants';
+import CheckBox from '../CheckBox';
+import Caption from '../Caption';
+import FilterInfo from '../FilterInfo';
+import RadioButton from '../RadioButton';
 
 class Choices extends Component {
     renderChoices() {
