+++ src/components/FullText/index.js
@@ -1,0 +1,1 @@
+export { default } from './FullText';
