+++ src/components/GroupTopLabel/index.js
@@ -1,0 +1,1 @@
+export { default } from './GroupTopLabel';
